requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2020.12.5
cffi==1.15.1
chardet==4.0.0
crispy-bootstrap5==0.7
cryptography==3.3.1
defusedxml==0.7.0rc1
Django==4.2
django-bootstrap-v5==1.0.11
django-ckeditor==6.0.0
django-crispy-forms==2.0
django-js-asset==1.2.2
django-widget-tweaks==1.4.8
idna==2.10
oauthlib==3.1.0
Pillow==9.5.0
pycparser==2.20
PyJWT==2.0.0
python-decouple==3.3
python3-openid==3.2.0
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
soupsieve==2.4.1
sqlparse==0.4.1
typing_extensions==4.6.3
urllib3==1.26.2
